requirements.txt
# Основная библиотека для генерации облаков слов
wordcloud>=1.9.0

# Основные библиотеки для работы с изображениями
Pillow>=9.0.0
numpy>=1.21.0

# Дополнительные библиотеки для расширенных возможностей
matplotlib>=3.5.0

# Для работы с шрифтами (опционально)
fonttools>=4.0.0

# Для обработки текста
regex>=2022.0.0

# Дополнительные зависимости для wordcloud
scipy>=1.7.0

#Бот
aiogram>=3.0.0
schedule>=1.2.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
